refactor: RacingGame를 Domain 패키지로 이동
