+++ src/main/java/racing/core/GameRunner.java
@@ -1,5 +1,6 @@
 package racing.core;
 
+import racing.core.domain.RacingGame;
 import racing.core.dto.Trial;
 import racing.core.patterns.RandomMoveStrategy;
 import racing.view.InputView;
