+++ src/test/java/racing/core/domain/RacingGameTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import racing.core.RacingGame;
 import racing.core.dto.Trial;
 import racing.core.exception.ErrorMessage;
